refactor: Moved type hints to its own module
