+++ cognite_toolkit/cdf_tk/utils.py
@@ -15,19 +15,16 @@ from __future__ import annotations
 
 import abc
 import collections
-import importlib
 import inspect
 import json
 import logging
 import os
-import re
 import typing
-from collections import Counter
 from collections.abc import Sequence
 from dataclasses import dataclass
 from functools import total_ordering
 from pathlib import Path
-from typing import Any, get_origin, get_type_hints
+from typing import Any, get_origin
 
 import yaml
 from cognite.client import ClientConfig, CogniteClient
@@ -36,7 +33,10 @@ from cognite.client.credentials import OAuthClientCredentials, Token
 from cognite.client.data_classes._base import CogniteObject
 from cognite.client.data_classes.capabilities import Capability
 from cognite.client.exceptions import CogniteAPIError, CogniteAuthError
+from cognite.client.utils._text import to_camel_case, to_snake_case
 from rich import print
+
+from cognite_toolkit.cdf_tk._get_type_hints import _TypeHints
 
 logger = logging.getLogger(__name__)
 
@@ -489,234 +489,3 @@ def _validate_case_raw(
                     )
 
     return warnings
-
-
-def to_camel_case(string: str) -> str:
-    """Convert snake_case_name to camelCaseName.
-
-    Args:
-        string: The string to convert.
-    Returns:
-        camelCase of the input string.
-
-    Examples:
-        >>> to_camel_case("a_b")
-        'aB'
-        >>> to_camel_case('camel_case')
-        'camelCase'
-        >>> to_camel_case('best_director')
-        'bestDirector'
-        >>> to_camel_case("ScenarioInstance_priceForecast")
-        'scenarioInstancePriceForecast'
-    """
-    if "_" in string:
-        # Could be a combination of snake and pascal/camel case
-        parts = string.split("_")
-        pascal_splits = [to_pascal_case(part) for part in parts]
-    else:
-        # Assume is pascal/camel case
-        # Ensure pascal
-        string = string[0].upper() + string[1:]
-        pascal_splits = [string]
-    string_split = []
-    for part in pascal_splits:
-        string_split.extend(re.findall(r"[A-Z][a-z]*", part))
-    if not string_split:
-        string_split = [string]
-    try:
-        return string_split[0].casefold() + "".join(word.capitalize() for word in string_split[1:])
-    except IndexError:
-        return ""
-
-
-def to_pascal_case(string: str) -> str:
-    """Convert string to PascalCaseName.
-
-    Args:
-        string: The string to convert.
-    Returns:
-        PascalCase of the input string.
-
-    Examples:
-        >>> to_pascal_case("a_b")
-        'AB'
-        >>> to_pascal_case('camel_case')
-        'CamelCase'
-        >>> to_pascal_case('best_director')
-        'BestDirector'
-        >>> to_pascal_case("ScenarioInstance_priceForecast")
-        'ScenarioInstancePriceForecast'
-    """
-    camel = to_camel_case(string)
-    return f"{camel[0].upper()}{camel[1:]}" if camel else ""
-
-
-def to_snake_case(string: str) -> str:
-    """
-    Convert input string to snake_case
-
-    Args:
-        string: The string to convert.
-    Returns:
-        snake_case of the input string.
-
-    Examples:
-        >>> to_snake_case("aB")
-        'a_b'
-        >>> to_snake_case('CamelCase')
-        'camel_case'
-        >>> to_snake_case('camelCamelCase')
-        'camel_camel_case'
-        >>> to_snake_case('Camel2Camel2Case')
-        'camel_2_camel_2_case'
-        >>> to_snake_case('getHTTPResponseCode')
-        'get_http_response_code'
-        >>> to_snake_case('get200HTTPResponseCode')
-        'get_200_http_response_code'
-        >>> to_snake_case('getHTTP200ResponseCode')
-        'get_http_200_response_code'
-        >>> to_snake_case('HTTPResponseCode')
-        'http_response_code'
-        >>> to_snake_case('ResponseHTTP')
-        'response_http'
-        >>> to_snake_case('ResponseHTTP2')
-        'response_http_2'
-        >>> to_snake_case('Fun?!awesome')
-        'fun_awesome'
-        >>> to_snake_case('Fun?!Awesome')
-        'fun_awesome'
-        >>> to_snake_case('10CoolDudes')
-        '10_cool_dudes'
-        >>> to_snake_case('20coolDudes')
-        '20_cool_dudes'
-    """
-    pattern = re.compile(r"[A-Z]?[a-z]+|[A-Z]+(?=[A-Z][a-z]|\d|\W|$)|\d+")
-    if "_" in string:
-        words = [word for section in string.split("_") for word in pattern.findall(section)]
-    else:
-        words = pattern.findall(string)
-    return "_".join(map(str.lower, words))
-
-
-class _TypeHints:
-    """
-    This class is used to get type hints from the init function of a CogniteObject.
-
-    After Python 3.10, type hints are treated as strings, so we need to evaluate them to get the actual type.
-    """
-
-    @classmethod
-    def get_type_hints_by_name(cls, signature, resource_cls: type[CogniteObject]) -> dict[str, Any]:
-        """
-        Get type hints from the init function of a CogniteObject.
-
-        Args:
-            signature: The signature of the init function.
-            resource_cls: The resource class to get type hints from.
-        """
-        try:
-            type_hint_by_name = get_type_hints(resource_cls.__init__, localns=cls._type_checking)
-        except TypeError:
-            # Python 3.10 Type hints cannot be evaluated with get_type_hints,
-            # ref https://stackoverflow.com/questions/66006087/how-to-use-typing-get-type-hints-with-pep585-in-python3-8
-            resource_module_vars = vars(importlib.import_module(resource_cls.__module__))
-            resource_module_vars.update(cls._type_checking())
-            type_hint_by_name = cls._get_type_hints_3_10(resource_module_vars, signature, vars(resource_cls))
-        return type_hint_by_name
-
-    @classmethod
-    def _type_checking(cls) -> dict[str, Any]:
-        """
-        When calling the get_type_hints function, it imports the module with the function TYPE_CHECKING is set to False.
-
-        This function takes all the special types used in data classes and returns them as a dictionary so it
-        can be used in the local namespaces.
-        """
-        import numpy as np
-        import numpy.typing as npt
-        from cognite.client import CogniteClient
-
-        NumpyDatetime64NSArray = npt.NDArray[np.datetime64]
-        NumpyInt64Array = npt.NDArray[np.int64]
-        NumpyFloat64Array = npt.NDArray[np.float64]
-        NumpyObjArray = npt.NDArray[np.object_]
-        return {
-            "CogniteClient": CogniteClient,
-            "NumpyDatetime64NSArray": NumpyDatetime64NSArray,
-            "NumpyInt64Array": NumpyInt64Array,
-            "NumpyFloat64Array": NumpyFloat64Array,
-            "NumpyObjArray": NumpyObjArray,
-        }
-
-    @classmethod
-    def _get_type_hints_3_10(
-        cls, resource_module_vars: dict[str, Any], signature310: inspect.Signature, local_vars: dict[str, Any]
-    ) -> dict[str, Any]:
-        return {
-            name: cls._create_type_hint_3_10(parameter.annotation, resource_module_vars, local_vars)
-            for name, parameter in signature310.parameters.items()
-            if name != "self"
-        }
-
-    @classmethod
-    def _create_type_hint_3_10(
-        cls, annotation: str, resource_module_vars: dict[str, Any], local_vars: dict[str, Any]
-    ) -> Any:
-        if annotation.endswith(" | None"):
-            annotation = annotation[:-7]
-        try:
-            return eval(annotation, resource_module_vars, local_vars)
-        except TypeError:
-            # Python 3.10 Type Hint
-            return cls._type_hint_3_10_to_8(annotation, resource_module_vars, local_vars)
-
-    @classmethod
-    def _type_hint_3_10_to_8(
-        cls, annotation: str, resource_module_vars: dict[str, Any], local_vars: dict[str, Any]
-    ) -> Any:
-        if cls._is_vertical_union(annotation):
-            alternatives = [
-                cls._create_type_hint_3_10(a.strip(), resource_module_vars, local_vars) for a in annotation.split("|")
-            ]
-            return typing.Union[tuple(alternatives)]
-        elif annotation.startswith("dict[") and annotation.endswith("]"):
-            if Counter(annotation)[","] > 1:
-                key, rest = annotation[5:-1].split(",", 1)
-                return dict[key.strip(), cls._create_type_hint_3_10(rest.strip(), resource_module_vars, local_vars)]
-            key, value = annotation[5:-1].split(",")
-            return dict[
-                cls._create_type_hint_3_10(key.strip(), resource_module_vars, local_vars),
-                cls._create_type_hint_3_10(value.strip(), resource_module_vars, local_vars),
-            ]
-        elif annotation.startswith("Mapping[") and annotation.endswith("]"):
-            if Counter(annotation)[","] > 1:
-                key, rest = annotation[8:-1].split(",", 1)
-                return typing.Mapping[
-                    key.strip(), cls._create_type_hint_3_10(rest.strip(), resource_module_vars, local_vars)
-                ]
-            key, value = annotation[8:-1].split(",")
-            return typing.Mapping[
-                cls._create_type_hint_3_10(key.strip(), resource_module_vars, local_vars),
-                cls._create_type_hint_3_10(value.strip(), resource_module_vars, local_vars),
-            ]
-        elif annotation.startswith("Optional[") and annotation.endswith("]"):
-            return typing.Optional[cls._create_type_hint_3_10(annotation[9:-1], resource_module_vars, local_vars)]
-        elif annotation.startswith("list[") and annotation.endswith("]"):
-            return list[cls._create_type_hint_3_10(annotation[5:-1], resource_module_vars, local_vars)]
-        elif annotation.startswith("tuple[") and annotation.endswith("]"):
-            return tuple[cls._create_type_hint_3_10(annotation[6:-1], resource_module_vars, local_vars)]
-        elif annotation.startswith("typing.Sequence[") and annotation.endswith("]"):
-            # This is used in the Sequence data class file to avoid name collision
-            return typing.Sequence[cls._create_type_hint_3_10(annotation[16:-1], resource_module_vars, local_vars)]
-        raise NotImplementedError(f"Unsupported conversion of type hint {annotation!r}. {cls._error_msg}")
-
-    @classmethod
-    def _is_vertical_union(cls, annotation: str) -> bool:
-        if "|" not in annotation:
-            return False
-        parts = [p.strip() for p in annotation.split("|")]
-        for part in parts:
-            counts = Counter(part)
-            if counts["["] != counts["]"]:
-                return False
-        return True
